+++ assets/js/dashboard/stats/devices/index.js
@@ -1,17 +1,14 @@
 import React from 'react';
 import { Link } from 'react-router-dom'
 
-import numberFormatter from '../number-formatter'
-import Bar from './bar'
-import MoreLink from './more-link'
-import * as api from '../api'
+import Browsers from './browsers'
+import OperatingSystems from './operating-systems'
+import FadeIn from '../../fade-in'
+import numberFormatter from '../../number-formatter'
+import Bar from '../bar'
+import MoreLink from '../more-link'
+import * as api from '../../api'
 
-
-function FadeIn({show, children}) {
-  const className = show ? "fade-enter-active" : "fade-enter"
-
-  return <div className={className}>{children}</div>
-}
 
 const EXPLANATION = {
   'Mobile': 'up to 576px',
@@ -95,154 +92,6 @@ class ScreenSizes extends React.Component {
             <span>{ this.label() }</span>
           </div>
           { this.state.sizes && this.state.sizes.map(this.renderScreenSize.bind(this)) }
-        </React.Fragment>
-      )
-    } else {
-      return <div className="text-center mt-44 font-medium text-gray-500">No data yet</div>
-    }
-  }
-
-  render() {
-    return (
-      <React.Fragment>
-        { this.state.loading && <div className="loading mt-44 mx-auto"><div></div></div> }
-        <FadeIn show={!this.state.loading}>
-          { this.renderList() }
-        </FadeIn>
-      </React.Fragment>
-    )
-  }
-}
-
-class Browsers extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {loading: true}
-  }
-
-  componentDidMount() {
-    this.fetchBrowsers()
-    if (this.props.timer) this.props.timer.onTick(this.fetchBrowsers.bind(this))
-  }
-
-  componentDidUpdate(prevProps) {
-    if (this.props.query !== prevProps.query) {
-      this.setState({loading: true, browsers: null})
-      this.fetchBrowsers()
-    }
-  }
-
-  fetchBrowsers() {
-    api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/browsers`, this.props.query)
-      .then((res) => this.setState({loading: false, browsers: res}))
-  }
-
-  renderBrowser(browser) {
-    const query = new URLSearchParams(window.location.search)
-    query.set('browser', browser.name)
-
-    return (
-      <div className="flex items-center justify-between my-1 text-sm" key={browser.name}>
-        <div className="w-full h-8" style={{maxWidth: 'calc(100% - 6rem)'}}>
-          <Bar count={browser.count} all={this.state.browsers} bg="bg-green-50" />
-          <span className="flex px-2" style={{marginTop: '-26px'}} >
-            <Link className="block truncate hover:underline" to={{search: query.toString()}}>
-              {browser.name}
-            </Link>
-          </span>
-        </div>
-        <span className="font-medium">{numberFormatter(browser.count)} <span className="inline-block text-xs w-8 text-right">({browser.percentage}%)</span></span>
-      </div>
-    )
-  }
-
-  label() {
-    return this.props.query.period === 'realtime' ? 'Current visitors' : 'Visitors'
-  }
-
-  renderList() {
-    if (this.state.browsers && this.state.browsers.length > 0) {
-      return (
-        <React.Fragment>
-          <div className="flex items-center mt-3 mb-2 justify-between text-gray-500 text-xs font-bold tracking-wide">
-            <span>Browser</span>
-            <span>{ this.label() }</span>
-          </div>
-          { this.state.browsers && this.state.browsers.map(this.renderBrowser.bind(this)) }
-        </React.Fragment>
-      )
-    } else {
-      return <div className="text-center mt-44 font-medium text-gray-500">No data yet</div>
-    }
-  }
-
-  render() {
-    return (
-      <React.Fragment>
-        { this.state.loading && <div className="loading mt-44 mx-auto"><div></div></div> }
-        <FadeIn show={!this.state.loading}>
-          { this.renderList() }
-        </FadeIn>
-      </React.Fragment>
-    )
-  }
-}
-
-class OperatingSystems extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {loading: true}
-  }
-
-  componentDidMount() {
-    this.fetchOperatingSystems()
-    if (this.props.timer) this.props.timer.onTick(this.fetchOperatingSystems.bind(this))
-  }
-
-  componentDidUpdate(prevProps) {
-    if (this.props.query !== prevProps.query) {
-      this.setState({loading: true, operatingSystems: null})
-      this.fetchOperatingSystems()
-    }
-  }
-
-  fetchOperatingSystems() {
-    api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/operating-systems`, this.props.query)
-      .then((res) => this.setState({loading: false, operatingSystems: res}))
-  }
-
-  renderOperatingSystem(os) {
-    const query = new URLSearchParams(window.location.search)
-    query.set('os', os.name)
-
-    return (
-      <div className="flex items-center justify-between my-1 text-sm" key={os.name}>
-        <div className="w-full h-8" style={{maxWidth: 'calc(100% - 6rem)'}}>
-          <Bar count={os.count} all={this.state.operatingSystems} bg="bg-green-50" />
-          <span className="flex px-2" style={{marginTop: '-26px'}}>
-            <Link className="block truncate hover:underline" to={{search: query.toString()}}>
-              {os.name}
-            </Link>
-          </span>
-        </div>
-        <span className="font-medium">{numberFormatter(os.count)} <span className="inline-block text-xs w-8 text-right">({os.percentage}%)</span></span>
-      </div>
-    )
-  }
-
-  label() {
-    return this.props.query.period === 'realtime' ? 'Current visitors' : 'Visitors'
-  }
-
-  renderList() {
-    if (this.state.operatingSystems && this.state.operatingSystems.length > 0) {
-      return (
-        <React.Fragment>
-          <div className="flex items-center mt-3 mb-2 justify-between text-gray-500 text-xs font-bold tracking-wide">
-            <span>Operating system</span>
-            <span>{ this.label() }</span>
-          </div>
-          { this.state.operatingSystems && this.state.operatingSystems.map(this.renderOperatingSystem.bind(this)) }
         </React.Fragment>
       )
     } else {