+++ lib/plausible_web/router.ex
@@ -51,7 +51,9 @@ defmodule PlausibleWeb.Router do
     get "/:domain/entry-pages", StatsController, :entry_pages
     get "/:domain/countries", StatsController, :countries
     get "/:domain/browsers", StatsController, :browsers
+    get "/:domain/browser-versions", StatsController, :browser_versions
     get "/:domain/operating-systems", StatsController, :operating_systems
+    get "/:domain/operating-system-versions", StatsController, :operating_system_versions
     get "/:domain/screen-sizes", StatsController, :screen_sizes
     get "/:domain/conversions", StatsController, :conversions
     get "/:domain/property/:prop_name", StatsController, :prop_breakdown
@@ -143,7 +145,13 @@ defmodule PlausibleWeb.Router do
 
     get "/:website/snippet", SiteController, :add_snippet
     get "/:website/settings", SiteController, :settings
-    get "/:website/goals", SiteController, :goals
+    get "/:website/settings/general", SiteController, :settings_general
+    get "/:website/settings/visibility", SiteController, :settings_visibility
+    get "/:website/settings/goals", SiteController, :settings_goals
+    get "/:website/settings/search-console", SiteController, :settings_search_console
+    get "/:website/settings/email-reports", SiteController, :settings_email_reports
+    get "/:website/settings/custom-domain", SiteController, :settings_custom_domain
+    get "/:website/settings/danger-zone", SiteController, :settings_danger_zone
     get "/:website/goals/new", SiteController, :new_goal
     post "/:website/goals", SiteController, :create_goal
     delete "/:website/goals/:id", SiteController, :delete_goal