+++ test/plausible_web/controllers/api/external_controller_test.exs
@@ -101,6 +101,21 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       assert get_event("external-controller-test-5.com") == nil
     end
 
+    test "Headless Chrome is ignored", %{conn: conn} do
+      params = %{
+        name: "pageview",
+        url: "http://www.example.com/",
+        domain: "headless-chrome-test.com"
+      }
+
+      conn
+      |> put_req_header("content-type", "text/plain")
+      |> put_req_header("user-agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/85.0.4183.83 Safari/537.36")
+      |> post("/api/event", Jason.encode!(params))
+
+      assert get_event("headless-chrome-test.com") == nil
+    end
+
     test "parses user_agent", %{conn: conn} do
       params = %{
         name: "pageview",
@@ -118,7 +133,9 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
 
       assert response(conn, 202) == ""
       assert pageview.operating_system == "Mac"
+      assert pageview.operating_system_version == "10.13"
       assert pageview.browser == "Chrome"
+      assert pageview.browser_version == "70.0"
     end
 
     test "parses referrer", %{conn: conn} do
@@ -374,6 +391,27 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
     assert event.name == "custom event"
   end
 
+  test "casts custom props to string", %{conn: conn} do
+    params = %{
+      name: "Signup",
+      url: "http://gigride.live/",
+      domain: "custom-prop-test.com",
+      props: Jason.encode!(%{
+        bool_test: true,
+        number_test: 12
+      })
+    }
+
+    conn
+    |> put_req_header("content-type", "text/plain")
+    |> post("/api/event", Jason.encode!(params))
+
+    event = get_event("custom-prop-test.com")
+
+    assert Map.get(event, :"meta.key") == ["bool_test", "number_test"]
+    assert Map.get(event, :"meta.value") == ["true", "12"]
+  end
+
   test "ignores a malformed referrer URL", %{conn: conn} do
     params = %{
       name: "pageview",
@@ -465,6 +503,25 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
     pageview = get_event("external-controller-test-23.com")
 
     assert pageview.pathname == "/#page-a"
+  end
+
+  test "decodes URL pathname, fragment and search", %{conn: conn} do
+    params = %{
+      n: "pageview",
+      u: "https://test.com/%EF%BA%9D%EF%BB%AD%EF%BA%8E%EF%BA%8B%EF%BA%AF-%EF%BB%AE%EF%BB%A4%EF%BA%B3%EF%BA%8E%EF%BA%92%EF%BB%97%EF%BA%8E%EF%BA%97?utm_source=%25balle%25",
+      d: "url-decode-test.com",
+      h: 1
+    }
+
+    conn
+    |> put_req_header("content-type", "text/plain")
+    |> post("/api/event", Jason.encode!(params))
+
+    pageview = get_event("url-decode-test.com")
+
+    assert pageview.hostname == "test.com"
+    assert pageview.pathname == "/ﺝﻭﺎﺋﺯ-ﻮﻤﺳﺎﺒﻗﺎﺗ"
+    assert pageview.utm_source == "%balle%"
   end
 
   test "responds 400 when required fields are missing", %{conn: conn} do