+++ test/plausible_web/controllers/api/stats_controller/operating_systems_test.exs
@@ -14,4 +14,17 @@ defmodule PlausibleWeb.Api.StatsController.OperatingSystemsTest do
              ]
     end
   end
+
+  describe "GET /api/stats/:domain/operating-system-versions" do
+    setup [:create_user, :log_in, :create_site]
+
+    test "returns top OS versions by unique visitors", %{conn: conn, site: site} do
+      filters = Jason.encode!(%{os: "Mac"})
+      conn = get(conn, "/api/stats/#{site.domain}/operating-system-versions?period=day&date=2019-01-01&filters=#{filters}")
+
+      assert json_response(conn, 200) == [
+               %{"name" => "10.15", "count" => 1, "percentage" => 100}
+             ]
+    end
+  end
 end